+++ flow/ui-admin/src/main/java/br/net/mirante/singular/service/FlowAuthorizationFacade.java
@@ -39,7 +39,7 @@ public class FlowAuthorizationFacade {
         return listProcessDefinitionKeysWithAccess(groupDTO, userCod, accessLevel).stream().map(definitionDAO::retrieveByKey).collect(Collectors.toSet());
     }
 
-    public Set<Long> listProcessDefinitionCodsWithAccess(GroupDTO groupDTO, String userCod, AccessLevel accessLevel) {
+    public Set<Integer> listProcessDefinitionCodsWithAccess(GroupDTO groupDTO, String userCod, AccessLevel accessLevel) {
         return listProcessDefinitionsWithAccess(groupDTO, userCod, accessLevel).stream().map(DefinitionDTO::getCod).collect(Collectors.toSet());
     }
